Split workspace_management tools into one file per tool
Follows the layout of the newer servers (files, data_sources_file_system):
each tool exports its handler from its own file, tools/index.ts only maps
names to handlers, and the shared pagination/rendering helpers move to
tools/utils.ts. No behaviour change.
